requirements.txt
# Python 3.11+ required
numpy>=1.26.0
Pillow>=10.2.0
pynput>=1.7.6
anthropic>=0.7.0
pyautogui>=0.9.54
configparser>=6.0.1
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
mss>=9.0.1
pywin32>=306
opencv-python>=4.9.0.80 